requirements.txt
fastapi
uvicorn
google.generativeai
python-dotenv
librosa
openai-whisper
scipy
sounddevice
ipywidgets
google-generativeai
redis
Click
h11